msm: kgsl: Use timer instead of loop count to wait gpu to idle
Previously kgsl_yamato_idle() used a loop count to check gpu hang.
This count is not big enough for the test app so a gpu hang was
reported too soon. A drawback of using loop constant is inconsistent
waiting time for different cpu speeds. To fix this, read cpu time to
check timeout. Also leave more time to wait for gpu to finish its work.